Former-commit-id: 6022549dfde93299e54a189c3d4da8e0175b7e41

+++ java/src/xmipp/viewer/particlepicker/ParticlePicker.java
@@ -327,7 +327,7 @@ public abstract class ParticlePicker
 					
 				}
 				else
-					family = new Family(name, new Color(rgb), size, state, this, templatesNumber.intValue());
+					family = new Family(name, new Color(rgb), size, state, this, templatesNumber);
 				families.add(family);
 			}
 			md.destroy();
@@ -459,6 +459,7 @@ public abstract class ParticlePicker
 				return m;
 		return null;
 	}
+
